requirements.txt
langchain
openai
spotipy
python-dotenv